fix(test): fix tests to work with new db

+++ database/migrations/2019_05_28_214118_add_user_id_to_transactions_table.php
@@ -27,8 +27,12 @@ class AddUserIdToTransactionsTable extends Migration
     public function down()
     {
         Schema::table('transactions', function (Blueprint $table) {
-            $table->dropColumn('from_user_id');
-            $table->dropColumn('to_user_id');
+            if (Schema::hasColumn('transactions', 'from_user_id')) {
+                $table->dropColumn('from_user_id');
+            }
+            if (Schema::hasColumn('transactions', 'to_user_id')) {
+                $table->dropColumn('to_user_id');
+            }
         });
     }
 }
